add lstm layer
