+++ zenu-layer/src/layers/rnn/mod.rs
@@ -1,8 +1,6 @@
 mod builder;
+mod inner;
 mod lstm;
-mod neo_lstm;
-mod neo_rnn;
-mod neo_struct;
 mod rnn;
 
 pub use lstm::*;
